% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.R
\docType{package}
\name{BKmisc-package}
\alias{BKmisc}
\alias{BKmisc-package}
\title{BKmisc: Variuos usefull function written by Bartlomiej Karaban}
\description{
@docType package
 @name BKmisc
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/karabanb/BKmisc}
  \item Report bugs at \url{https://github.com/karabanb/BKmisc/issues}
}

}
