% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_rpart.R
\name{discretize_rpart}
\alias{discretize_rpart}
\title{Dicretizing Continous Values by Decision Tree}
\usage{
discretize_rpart(x, y, maxdepth = 2, minbucket = 10)
}
\arguments{
\item{x}{- data.frame with continues variables}

\item{y}{- Target binary variable}

\item{maxdepth}{- parameter from rpart::rpart.control. It sets number of expected bins.}

\item{minbucket}{- parameter from rpart::rpart.control. It sets minumum size of bin.}
}
\value{

}
\description{
Dicretizing Continous Values by Decision Tree
}
\examples{

}
