% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_vars.R
\name{email_vars}
\alias{email_vars}
\title{Title Extract features form email address, name and surname}
\usage{
email_vars(x, address, name, surname)
}
\arguments{
\item{x}{- a data.frame object with email address, name and surname.}

\item{address}{- column's name with email address}

\item{name}{- column's name with email's owner name}

\item{surname}{- column's name with of email's owner surname}
}
\value{

}
\description{
Title Extract features form email address, name and surname
}
\examples{

}
