% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_unique_dt.R
\name{select_unique_dt}
\alias{select_unique_dt}
\title{Select variables which contain at least two unique values by \code{data.table} way}
\usage{
select_unique_dt(x)
}
\arguments{
\item{x}{data.frame or matrix - data for cleaning}
}
\value{
This function return a \code{data.frame} or \code{matrix} which contains columns with at leat two uniques
values.
}
\description{
Select variables which contain at least two unique values by \code{data.table} way
}
\examples{
df <- data.frame(v1 = 1:4, v2 = 1, v3 = LETTERS[1:4], v4 = c(1, 1, NA, NA))
df_unq <- select_unique(df)
df; df_unq

}
\author{
Bartlomiej Karaban
}
