% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_unique_dt.R
\name{select_unique_dt}
\alias{select_unique_dt}
\title{Select variables which contain at least two unique values by \code{data.table} way}
\usage{
select_unique_dt(x)
}
\arguments{
\item{x}{data.frame or matrix - data for cleaning}
}
\value{
This function return a \code{data.frame} or \code{matrix} which contains columns with at leat two uniques
values.
}
\description{
Select variables which contain at least two unique values by \code{data.table} way
}
\author{
Bartlomiej Karaban
}
